tgbot/requirements.txt
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==4.4.2
google-api-core==2.19.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
moviepy==1.0.3
numpy==1.26.4
oauthlib==3.2.2
pillow==10.3.0
proglog==0.1.10
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
pyTelegramBotAPI==4.18.1
python-dotenv==1.0.1
pytube==15.0.0
requests==2.32.2
requests-oauthlib==2.0.0
rsa==4.9
setuptools==70.0.0
tqdm==4.66.4
uritemplate==4.1.1
urllib3==2.2.1
validators==0.28.3
